requirements.txt
click~=7.1.2
matplotlib~=3.4.2
numpy~=1.20.3
pandas~=1.2.4
parameterized~=0.8.1
Pillow~=8.2.0
pytorch-lightning~=1.3.0
pytorch-metric-learning~=0.9.99
pytorch-nlp~=0.5.0
requests~=2.25.1
six~=1.16.0
scipy~=1.6.3
scikit-learn~=0.24.2
setuptools~=56.2.0
sphinx~=4.0.1
sqlalchemy~=1.4.15
tensorboard~=2.5.0
texttable~=1.6.3
# torch~=1.8.1
# torch~=1.8.1+cpu
# torchaudio~=0.8.1
torchinfo~=0.1.1
# torchtext~=0.9.1
# torchvision~=0.9.1
# torchvision~=0.9.1+cpu
tqdm~=4.60.0
umap-learn~=0.5.1
